# openapi.yaml
openapi: "3.0.3"
info:
  title: Grunt API
  version: 0.2.0
  description: Real-time messaging server API

servers:
  - url: http://localhost:54765

paths:
  /api/user:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/user/invite:
    get:
      summary: Generate a new invite code
      operationId: getInvite
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Invite code generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/user/login:
    post:
      summary: Authenticate and obtain a token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/chat/message:
    post:
      summary: Send a message
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/chat/stream:
    get:
      summary: Stream messages and events via SSE
      operationId: streamMessages
      security:
        - BearerAuth: []
      responses:
        '200':
          description: SSE stream of messages and events
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/chat/sync:
    get:
      summary: Sync messages
      operationId: syncMessages
      parameters:
        - name: since
          in: query
          description: Return messages with ID greater than this value
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: last
          in: query
          description: Maximum number of messages to return (used when since=0)
          required: false
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Broadcast'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/users:
    post:
      summary: Create a new user (admin only)
      operationId: adminCreateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/api-keys:
    get:
      summary: List all API keys (admin only)
      operationId: adminListAPIKeys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyInfoFull'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new API key (admin only)
      operationId: adminCreateAPIKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateAPIKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/api-keys/{keyId}:
    delete:
      summary: Revoke an API key (admin only)
      operationId: adminRevokeAPIKey
      security:
        - BearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - user
        - password
        - invite_code
      properties:
        user:
          type: string
          description: Username
          example: alice
        password:
          type: string
          description: Password
          example: secret123
        invite_code:
          type: string
          description: Invite code (required for registration)
          example: abc123

    InviteResponse:
      type: object
      properties:
        code:
          type: string
          description: Invite code
          example: abc123
        expires_at:
          type: string
          format: date-time
          description: Expiration time
          example: 2026-05-07T00:10:00Z

    LoginRequest:
      type: object
      required:
        - user
        - password
      properties:
        user:
          type: string
          description: Username
          example: alice
        password:
          type: string
          description: Password
          example: secret123

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Authentication token (valid for 24 hours)
          example: abc123def456

    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Message content
          example: Hello, world!

    SendMessageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Message ID
          example: 42

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: user created

    Broadcast:
      type: object
      properties:
        type:
          type: string
          example: message
        id:
          type: integer
          format: int64
          example: 42
        content:
          type: string
          example: Hello, world!
        client_id:
          type: string
          example: abc123
        user:
          type: string
          example: alice
        timestamp:
          type: string
          format: date-time
          example: 2026-05-07T00:00:00Z

    Event:
      type: object
      properties:
        type:
          type: string
          example: event
        event:
          type: string
          enum:
            - join
            - leave
          example: join
        client_id:
          type: string
          example: abc123
        user:
          type: string
          example: alice

    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid credentials

    AdminCreateUserRequest:
      type: object
      required:
        - user
      properties:
        user:
          type: string
          description: Username
          example: bot-user

    AdminCreateAPIKeyRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          description: The user ID to create the API key for
          example: igor
        name:
          type: string
          description: Optional label for the key
          example: "igor-bot"

    APIKeyInfoFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Key ID
          example: 1
        user_id:
          type: string
          description: User ID the key belongs to
          example: igor
        name:
          type: string
          description: Optional label
          example: "igor-bot"
        created_at:
          type: string
          format: date-time
          description: Creation time
          example: 2026-05-07T00:00:00Z

    APIKeyResponse:
      type: object
      properties:
        key_id:
          type: integer
          format: int64
          description: Key ID
          example: 1
        secret:
          type: string
          description: The raw API key (shown only once)
          example: gk_abc123def456